% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_convert.R
\name{trans_convert}
\alias{trans_convert}
\title{Abundance table transformation.}
\description{
Abundance table standardization, transformation or normalization for the otu_table in microtable object.
The input dataset must be a microtable object.
}
\examples{

## ------------------------------------------------
## Method `trans_convert$new`
## ------------------------------------------------

library(microeco)
data(dataset)
t1 <- trans_convert$new(dataset = dataset)

## ------------------------------------------------
## Method `trans_convert$convert`
## ------------------------------------------------

newdataset <- t1$convert(method = "log")
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-trans_convert-new}{\code{trans_convert$new()}}
\item \href{#method-trans_convert-convert}{\code{trans_convert$convert()}}
\item \href{#method-trans_convert-clone}{\code{trans_convert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_convert-new"></a>}}
\if{latex}{\out{\hypertarget{method-trans_convert-new}{}}}
\subsection{Method \code{new()}}{
Get a transposed abundance table in the object. Rows are samples and columns are features. 
This can make the further operations same with the traditional ecological methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_convert$new(dataset = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} Class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
transposed abundance table stored in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{library(microeco)
data(dataset)
t1 <- trans_convert$new(dataset = dataset)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_convert-convert"></a>}}
\if{latex}{\out{\hypertarget{method-trans_convert-convert}{}}}
\subsection{Method \code{convert()}}{
The function offers some methods based on the \code{\link{decostand}} function in vegan package and several famous methods in microbial field.
MARGIN = 1 to represent samples, i.e. rows and MARGIN = 2 to represent features, i.e. columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_convert$convert(method = NULL, MARGIN = NULL, logbase = exp(1), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{default NULL; See the following details and available options. \cr 
\cr 
Methods from \code{\link{decostand}}:
\itemize{
  \item \code{total}: divide by margin total (default MARGIN = 1, i.e. rows - samples).
  \item \code{max}: divide by margin maximum (default MARGIN = 2, i.e. columns - features).
  \item \code{normalize}:  make margin sum of squares equal to one (default MARGIN = 1).
  \item \code{range}: standardize values into range 0...1 (default MARGIN = 2). If all values are constant, they will be transformed to 0.
  \item \code{standardize}: scale x to zero mean and unit variance (default MARGIN = 2).
  \item \code{pa}: scale x to presence/absence scale (0/1).
  \item \code{log}: logarithmic transformation as suggested by Anderson et al. (2006): log_b (x) + 1 for x > 0, where b is the base of the logarithm; zeros are left as zeros. Higher bases give less weight to quantities and more to presences, and logbase = Inf gives the presence/absence scaling. Please note this is not log(x+1). Anderson et al. (2006) suggested this for their (strongly) modified Gower distance (implemented as method = "altGower" in vegdist), but the standardization can be used independently of distance indices.
}
Methods useful in microbial field:
\itemize{
  \item \code{CLR}: Centered log-ratio normalization (default MARGIN = 2, i.e. features).
}}

\item{\code{MARGIN}}{default NULL; 1 = samples, and 2 = features of abundance table.}

\item{\code{logbase}}{default exp(1); The logarithm base used in method = "log" or "CLR".}

\item{\code{...}}{parameters pass to \code{\link{decostand}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new microtable object; rows are features.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{newdataset <- t1$convert(method = "log")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_convert-clone"></a>}}
\if{latex}{\out{\hypertarget{method-trans_convert-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_convert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
