% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_gamma.R
\name{trans_gamma}
\alias{trans_gamma}
\title{Simulate gamma-beta diversity relationships under specific species distribution.}
\description{
Simulate gamma-beta diversity relationships under specific species distribution, and plot the result
based on Zhang et al. (2020). Local community assembly mechanisms shape soil bacterial β diversity patterns along a latitudinal gradient. Nat Commun 11, 5428 (2020).
<doi:10.1038/s41467-020-19228-4>.
}
\examples{

## ------------------------------------------------
## Method `trans_gamma$new`
## ------------------------------------------------

\donttest{
test1 <- trans_gamma$new(dataset = dataset, group = "Group")
}		

## ------------------------------------------------
## Method `trans_gamma$cal_observed`
## ------------------------------------------------

\donttest{
 test1$cal_observed(sample_size = NULL)
}

## ------------------------------------------------
## Method `trans_gamma$plot_observed`
## ------------------------------------------------

\donttest{
test1$plot_observed(cor_method = "spearman")
}

## ------------------------------------------------
## Method `trans_gamma$cal_simulation`
## ------------------------------------------------

\donttest{
test1 <- t1$cal_simulation(ncom =20)
}

## ------------------------------------------------
## Method `trans_gamma$plot_simulation`
## ------------------------------------------------

\donttest{
test1$plot_simulation()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_gamma$new()}}
\item \href{#method-cal_observed}{\code{trans_gamma$cal_observed()}}
\item \href{#method-plot_observed}{\code{trans_gamma$plot_observed()}}
\item \href{#method-cal_simulation}{\code{trans_gamma$cal_simulation()}}
\item \href{#method-plot_simulation}{\code{trans_gamma$plot_simulation()}}
\item \href{#method-print}{\code{trans_gamma$print()}}
\item \href{#method-clone}{\code{trans_gamma$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This function is used to simulate gamma-beta diversity relationships under specific species distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$new(dataset = NULL, group = NULL, method = "bray", seed = 123)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{default NULL; microtable object; used for the observed pattern.}

\item{\code{group}}{default NULL; a column name in sample_table; used as the different species pool for groups.}

\item{\code{method}}{default "bray"; dissimilarity indices; see \code{\link{vegdist}} function and method parameter in vegan package.}

\item{\code{seed}}{default 123; random seed used for the fixed random number generator for the repeatability.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
parameters in object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1 <- trans_gamma$new(dataset = dataset, group = "Group")
}		
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_observed"></a>}}
\if{latex}{\out{\hypertarget{method-cal_observed}{}}}
\subsection{Method \code{cal_observed()}}{
Calculate observed gamma and beta diversity for each group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$cal_observed(sample_size = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_size}}{default NULL; a numeric vector for the rarefied and uniform individual numbers in each sample; 
If null, use the observed data; If provided, use the rarefied data, e.g. c(500, 2000).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
res_observed in object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
 test1$cal_observed(sample_size = NULL)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_observed"></a>}}
\if{latex}{\out{\hypertarget{method-plot_observed}{}}}
\subsection{Method \code{plot_observed()}}{
Plot the observed result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$plot_observed(
  x_axis_title = "Gamma diversity",
  y_axis_title = "Beta diversity",
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_axis_title}}{default "Gamma diversity"; x axis title.}

\item{\code{y_axis_title}}{default "Beta diversity"; y axis title.}

\item{\code{...}}{parameters pass to the function plot_scatterfit in \code{\link{trans_env}} Class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1$plot_observed(cor_method = "spearman")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cal_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-cal_simulation}{}}}
\subsection{Method \code{cal_simulation()}}{
Simulate gamma-beta diversity relationships under log-normal abundance distribution without any measured data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$cal_simulation(
  gamma_vect = seq(1, 10000, by = 200),
  ind_vect = c(500, 1500, 3000, 5500, 10000),
  ncom = 100
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{gamma_vect}}{default seq(1, 10000, by = 200); a vector as gamma diversity.}

\item{\code{ind_vect}}{default c(500, 1500, 3000, 5500, 10000); a vector as individuals per plot.}

\item{\code{ncom}}{default 100; number of communities; how many communities or samples in the region or the studied species pool.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
res_simulation in object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1 <- t1$cal_simulation(ncom =20)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-plot_simulation}{}}}
\subsection{Method \code{plot_simulation()}}{
Plot the simulation result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$plot_simulation(
  color_values = RColorBrewer::brewer.pal(8, "Dark2"),
  show_point = TRUE,
  point_size = 0.3,
  point_alpha = 0.6,
  add_fitting = FALSE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{color_values}}{colors used for presentation.}

\item{\code{show_point}}{default TRUE; whether show the point.}

\item{\code{point_size}}{default .3; point size value.}

\item{\code{point_alpha}}{default .6; point alpha value.}

\item{\code{add_fitting}}{default FALSE; whether add fitted line.}

\item{\code{...}}{parameters pass to ggplot2::geom_line (when add_fitting = FALSE) or ggplot2::geom_smooth (when add_fitting = TRUE).}

\item{\code{color}}{default "SampleID"; color mapping in the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
ggplot.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1$plot_simulation()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the trans_gamma object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_gamma$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
