% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_netchord.R
\name{trans_netchord}
\alias{trans_netchord}
\title{Sum and plot the links number of the network.}
\description{
Sum and plot the links number from one taxa to another or in the same taxa in the network.
The input dataset must be a trans_network object.
}
\examples{

## ------------------------------------------------
## Method `trans_netchord$new`
## ------------------------------------------------

\donttest{
test1 <- trans_netchord$new(dataset = trans_network_object, taxa_level = "Phylum")
}

## ------------------------------------------------
## Method `trans_netchord$plot_sum_links`
## ------------------------------------------------

\donttest{
test1$plot_sum_links(plot_pos = TRUE, plot_num = 10)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{trans_netchord$new()}}
\item \href{#method-plot_sum_links}{\code{trans_netchord$plot_sum_links()}}
\item \href{#method-print}{\code{trans_netchord$print()}}
\item \href{#method-clone}{\code{trans_netchord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
This function is used to sum the links number from one taxa to another or in the same taxa, for example, at Phylum level.
This is very useful to fast see how many nodes are connected between different taxa or within the taxa.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_netchord$new(dataset = NULL, taxa_level = "Phylum")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{default NULL; trans_network object.}

\item{\code{taxa_level}}{default "Phylum"; taxonomic rank.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
res_sum_links_pos and res_sum_links_neg in object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1 <- trans_netchord$new(dataset = trans_network_object, taxa_level = "Phylum")
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_sum_links"></a>}}
\if{latex}{\out{\hypertarget{method-plot_sum_links}{}}}
\subsection{Method \code{plot_sum_links()}}{
Plot the summed linkages among taxa using chorddiag package <https://github.com/mattflor/chorddiag>.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_netchord$plot_sum_links(
  plot_pos = TRUE,
  plot_num = NULL,
  color_values = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_pos}}{default TRUE; plot the summed positive or negative linkages.}

\item{\code{plot_num}}{default NULL; number of taxa presented in the plot.}

\item{\code{color_values}}{default NULL; If not provided, use default.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
chorddiag plot
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\donttest{
test1$plot_sum_links(plot_pos = TRUE, plot_num = 10)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the trans_netchord object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_netchord$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_netchord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
