% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_microb_ts}
\alias{gut_microb_ts}
\title{The gut_microb_ts data}
\usage{
data(gut_microb_ts)
}
\description{
The gut_microb_ts data is a group of time series data used for the example of trans_ts class. 
The dataset comes from the study Gibbons et al. 2017. <doi: 10.1371/journal.pcbi.1005364>
}
\keyword{data.frame}
