% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_ts.R
\name{trans_ts}
\alias{trans_ts}
\title{Biomass evaluation and biological interaction prediction for Time Series Data analysis.}
\description{
Biomass evaluation and biological interaction prediction for Time Series Data analysis based on the method from BEEM package, Li et al. 2019. <doi: 10.1186/s40168-019-0729-z>.
}
\examples{

## ------------------------------------------------
## Method `trans_ts$new`
## ------------------------------------------------

data(gut_microb_ts)
t1 <- trans_ts$new(dataset = gut_microb_ts, filter_thres = 0)

## ------------------------------------------------
## Method `trans_ts$cal_biomass`
## ------------------------------------------------

t1$cal_biomass(min_iter = 50)

## ------------------------------------------------
## Method `trans_ts$cal_network`
## ------------------------------------------------

t1$t1$cal_network()
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-trans_ts-new}{\code{trans_ts$new()}}
\item \href{#method-trans_ts-cal_biomass}{\code{trans_ts$cal_biomass()}}
\item \href{#method-trans_ts-cal_network}{\code{trans_ts$cal_network()}}
\item \href{#method-trans_ts-clone}{\code{trans_ts$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_ts-new"></a>}}
\if{latex}{\out{\hypertarget{method-trans_ts-new}{}}}
\subsection{Method \code{new()}}{
Prepare data for the following analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_ts$new(dataset = NULL, filter_thres = 0.001)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{the object of \code{\link{microtable}} class.
Two columns with exact names in sample_table are necessary; one is 'Time', which is the time point and should be the numeric class;
the other is 'Rep', which represents the biological replicates and is also numeric class. If no replicates, use 1 to represent 1 replicate.}

\item{\code{filter_thres}}{default 0.001; the relative abundance threshold of taxa.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
abund_table and sample_table stored in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{data(gut_microb_ts)
t1 <- trans_ts$new(dataset = gut_microb_ts, filter_thres = 0)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_ts-cal_biomass"></a>}}
\if{latex}{\out{\hypertarget{method-trans_ts-cal_biomass}{}}}
\subsection{Method \code{cal_biomass()}}{
Predict the biomass.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_ts$cal_biomass(
  min_iter = 30,
  max_iter = 100,
  verbose = TRUE,
  scaling = 10000,
  seed = 1,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min_iter}}{minimal number of iterations for the EM algorithm (default: 30)}

\item{\code{max_iter}}{maximal number of iterations for the EM algorithm (default: 100)}

\item{\code{verbose}}{print more information (default: TRUE)}

\item{\code{scaling}}{median total biomass to scale all biomass data (default:10000)}

\item{\code{seed}}{random seed used in BLASSO (default:NULL)}

\item{\code{...}}{parameters passed to \code{\link{EM}} function in beem package.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
res_biomass and res_param stored in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t1$cal_biomass(min_iter = 50)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_ts-cal_network"></a>}}
\if{latex}{\out{\hypertarget{method-trans_ts-cal_network}{}}}
\subsection{Method \code{cal_network()}}{
Extract network result from EM results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_ts$cal_network(sig = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sig}}{significance threshold for the interaction evaluation (default: 0.05)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
res_network stored in the object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{t1$t1$cal_network()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-trans_ts-clone"></a>}}
\if{latex}{\out{\hypertarget{method-trans_ts-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{trans_ts$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
